render project links in index.html
At the moment, the project data is held in JSON and the list is created
in javascript after the page loads. This breaks the site for users
without JS.

This moves the rendering of the project list to index.html, so it's
included in the static page, and then the renderProjects() function
shows/hides and reorders the projects as necessary.

If the JS doesn't run, you just see the complete list of projects in the
default order, without the filter interface.