Create PlasmaPyWebsite.yml
Hi!

UpForGrabs has been awesome for our project's visibility. We're working on a website though, and we could use some help with that as well as we're completely in the dark about web development, basically. Do you think work with you guys on this one as well?

Thanks for everything!