Added Duality Docs Repo
Duality is a 2D game development framework that's [already part of upforgrabs](https://github.com/up-for-grabs/up-for-grabs.net/blob/gh-pages/_data/projects/duality.yml) for a few years. We recently split up our documentation, so it's now in a separate project, along with all issues related to the docs themselves. To get them listed again, I've added the docs project to upforgrabs separately, as we appreciate contributions to documentation as much as to the project core.

Let me know whether you think this makes sense for upforgrabs - not sure if tech writing should be a part of it, or whether the focus should be just coding itself, in which case just let me know and close the PR unmerged.