Updated default.html
I really enjoyed this, This is my first repo and i was nice solving this issue without a IDE giving me the answers. Some of the misplacement of elements, really made me have to think about it to since i'm used to coding from start to finish and not fixing code. Interesting skill i realize is diffrent than just "knowing the code".