Adding Scribe-iOS (#3271)
* Adding Scribe-iOS

Hi all 👋 Am hoping to get Scribe-iOS onto Up for Grabs :) I've been working on the project for a bit over half a year at this point, and slowly have been getting contributors. Scribe uses Wikidata as a backend, so I've been in active communication with many at Wikimedia about it, and it's  recently been added to the [featured projects for new Wikimedia developers](https://www.mediawiki.org/wiki/New_Developers).

Would be happy to hear your all's feedback on adding it, and hope the suggested `.yml` file meets the specifications!

* Update scribe-ios.yml issue link

Updating link to the one specified in review.