Add Together Java's TJ-Bot (#3334)
* Add Together Java's TJ-Bot

Hi!

TJ-Bot is a Discord Bot used on the Together Java server. It is maintained by the community, anyone can contribute!

The idea behind this project is to provide learning experience on a real-world project. Contributors will be provided with our guidance at every step of the way, and friendly but thorough reviews!

Looking forward to your feedback! :relaxed:

* Rename TJ-Bot to TJ-Bot.yml