fix for "popular tags" buttons not working
I noticed that the "popular tags" buttons did nothing when pressed.
I've amended a section of javascript that handles the popular tag elements, so that when you click on one, it adds it to the URL search query AND adds a visual element into the "tags" field too.

added contribution details too.