Fix click on popular tags
Fixes #5216

This PR restores the functionality of clicking on Popular Tags by correcting a bug in the popular tag parsing.
It also enables a second click to remove the tag from the filter list.

The existing code uses code to handle the project count element nested within the anchor containing the tag.
This new code is more conservative by extracting the tag from the first TEXT_NODE, to avoid making assumptions about other child elements which was necessary by using innerText.

Unlike #5217, this PR does not conflict with the existing event listener or the override already-filtered tags introduced by #1121 (and later fixed in #1179).